Add a way to check and dereference a pointer
Original merge commit: 31a3fae9f893565392e35fd3df1612615363dcdd

######

Add a way to check and dereference a pointer
It is common to get a pointer, check it is not null, and dereference it.
Sometimes that requires a named temporary just to be able to do the check.

The macro `DEREF(p)` provides this capability: it asserts that `p` is not null
and returns `*p`. This is analagous to `.value()` on an `std::optional`.

We might want to add a way to disable `CHECK` and the check in `DEREF` together.

This change also includes some examples of making use of `DEREF`.


Original commit: d7aa90e55ac80c7f2460ab7b0cb6d1ef0c068938
-------